# experiment: faster dev iterations with ext plugins