--- a/cli/build.rs
+++ b/cli/build.rs
@@ -340,7 +340,7 @@ fn create_cli_snapshot(snapshot_path: PathBuf) {
     ),
     deno_fetch::deno_fetch::init_ops::<PermissionsContainer>(Default::default()),
     deno_cache::deno_cache::init_ops::<SqliteBackedCache>(None),
-    deno_websocket::deno_websocket::init_ops::<PermissionsContainer>(
+    deno_websocket::deno_websocket::init_ops(
       "".to_owned(),
       None,
       None,