--- a/core/gotham_state.rs
+++ b/core/gotham_state.rs
@@ -3,6 +3,7 @@
 // https://github.com/gotham-rs/gotham/blob/bcbbf8923789e341b7a0e62c59909428ca4e22e2/gotham/src/state/mod.rs
 // Copyright 2017 Gotham Project Developers. MIT license.
 
+
 use log::trace;
 use std::any::type_name;
 use std::any::Any;
@@ -11,7 +12,20 @@ use std::collections::BTreeMap;
 
 #[derive(Default)]
 pub struct GothamState {
+  #[cfg(not(feature = "unstable_type_name"))]
   data: BTreeMap<TypeId, Box<dyn Any>>,
+  #[cfg(feature = "unstable_type_name")]
+  data: BTreeMap<&'static str, Box<dyn Any>>,
+}
+
+#[cfg(not(feature = "unstable_type_name"))]
+fn type_id<T: 'static>() -> TypeId {
+  TypeId::of::<T>()
+}
+
+#[cfg(feature = "unstable_type_name")]
+const fn type_id<T: 'static>() -> &'static str {
+  std::any::type_name::<T>()
 }
 
 impl GothamState {
@@ -19,20 +33,20 @@ impl GothamState {
   /// Successive calls to `put` will overwrite the existing value of the same
   /// type.
   pub fn put<T: 'static>(&mut self, t: T) {
-    let type_id = TypeId::of::<T>();
+    let type_id = type_id::<T>();
     trace!(" inserting record to state for type_id `{:?}`", type_id);
     self.data.insert(type_id, Box::new(t));
   }
 
   /// Determines if the current value exists in `GothamState` storage.
   pub fn has<T: 'static>(&self) -> bool {
-    let type_id = TypeId::of::<T>();
+    let type_id = type_id::<T>();
     self.data.get(&type_id).is_some()
   }
 
   /// Tries to borrow a value from the `GothamState` storage.
   pub fn try_borrow<T: 'static>(&self) -> Option<&T> {
-    let type_id = TypeId::of::<T>();
+    let type_id = type_id::<T>();
     trace!(" borrowing state data for type_id `{:?}`", type_id);
     self.data.get(&type_id).and_then(|b| b.downcast_ref())
   }
@@ -44,7 +58,7 @@ impl GothamState {
 
   /// Tries to mutably borrow a value from the `GothamState` storage.
   pub fn try_borrow_mut<T: 'static>(&mut self) -> Option<&mut T> {
-    let type_id = TypeId::of::<T>();
+    let type_id = type_id::<T>();
     trace!(" mutably borrowing state data for type_id `{:?}`", type_id);
     self.data.get_mut(&type_id).and_then(|b| b.downcast_mut())
   }
@@ -56,7 +70,7 @@ impl GothamState {
 
   /// Tries to move a value out of the `GothamState` storage and return ownership.
   pub fn try_take<T: 'static>(&mut self) -> Option<T> {
-    let type_id = TypeId::of::<T>();
+    let type_id = type_id::<T>();
     trace!(
       " taking ownership from state data for type_id `{:?}`",
       type_id