--- a/runtime/worker.rs
+++ b/runtime/worker.rs
@@ -218,7 +218,7 @@ impl MainWorker {
     fn dynamic<F>(
       name: &str,
     ) -> F {
-      let lib_path = format!("target/debug/lib{}.dylib", name);
+      let lib_path = format!("target/release/lib{}.dylib", name);
       let lib = dlopen::raw::Library::open(lib_path).unwrap();
       let fn_name = format!("dyn_init_{}", name);
       let f = unsafe {
@@ -229,9 +229,12 @@ impl MainWorker {
     }
 
     let ws = dynamic::<
-     fn() -> *mut deno_websocket::deno_websocket
+     fn(
+    user_agent: String,
+    root_cert_store_provider: Option<Arc<dyn RootCertStoreProvider>>,
+    unsafely_ignore_certificate_errors: Option<Vec<String>>
+         ) -> Extension
     >("deno_websocket");
-    let ws = unsafe { &mut *ws() };
 
     // NOTE(bartlomieju): ordering is important here, keep it in sync with
     // `runtime/build.rs`, `runtime/web_worker.rs` and `cli/build.rs`!
@@ -256,7 +259,7 @@ impl MainWorker {
         },
       ),
       deno_cache::deno_cache::init_ops::<SqliteBackedCache>(create_cache),
-      ws.init_ops_self::<PermissionsContainer>(
+      ws(
         options.bootstrap.user_agent.clone(),
         options.root_cert_store_provider.clone(),
         options.unsafely_ignore_certificate_errors.clone(),