@littledivy 2024-01-08
Solved by https://github.com/littledivy/plonk

Still need to figure out generics across crate boundaries.